+++ Nominas/frmReportes.cs
@@ -177,62 +177,77 @@ namespace Nominas
                 {
                     if (noReporte == 8)
                     {
+                        int existeNullCodeQR = 0;
                         cnx = new SqlConnection(cdn);
                         cmd = new SqlCommand();
                         cmd.Connection = cnx;
                         nh = new CalculoNomina.Core.NominaHelper();
                         nh.Command = cmd;
 
-                        List<CalculoNomina.Core.CodigoBidimensional> lstXml = new List<CalculoNomina.Core.CodigoBidimensional>();
                         try
                         {
                             cnx.Open();
-                            lstXml = nh.obtenerListaQr(GLOBALES.IDEMPRESA, dtpInicioPeriodo.Value.Date, dtpFinPeriodo.Value.Date);
+                            existeNullCodeQR = nh.existeNullQR(GLOBALES.IDEMPRESA, dtpInicioPeriodo.Value.Date, dtpFinPeriodo.Value.Date);
                             cnx.Close();
                         }
                         catch (Exception error)
                         {
-                            MessageBox.Show("Error: Al obtener el listado de los XML." + error.Message, "Error");
+                            MessageBox.Show("Error: Al obtener existencia de nulos Code QR." + error.Message, "Error");
                             cnx.Dispose();
                             return;
                         }
-                        
-                        string codigoQR = "";
-                        string[] valores = null;
-                        string numero = "";
-                        string vEntero = "";
-                        string vDecimal = "";
-                        
-                        for (int i = 0; i < lstXml.Count; i++)
+                        if (existeNullCodeQR != 0)
                         {
-                            numero = lstXml[i].tt.ToString();
-                            valores = numero.Split('.');
-                            vEntero = valores[0];
-                            vDecimal = valores[1];
-                            codigoQR = string.Format("?re={0}&rr={1}&tt={2}.{3}&id={4}",lstXml[i].re, lstXml[i].rr, 
-                                vEntero.PadLeft(10,'0'), vDecimal.PadRight(6, '0'), lstXml[i].uuid);
-                            var qrEncoder = new QrEncoder(ErrorCorrectionLevel.H);
-                            var qrCode = qrEncoder.Encode(codigoQR);
-                            var renderer = new GraphicsRenderer(new FixedModuleSize(2, QuietZoneModules.Two), Brushes.Black, Brushes.White);
-                            
-                            using (var stream = new FileStream(lstXml[i].uuid + ".png", FileMode.Create))
-                                renderer.WriteToStream(qrCode.Matrix, ImageFormat.Png, stream);
-
-                            Bitmap bmp = new Bitmap(lstXml[i].uuid + ".png");
-                            Byte[] qr = GLOBALES.IMAGEN_BYTES(bmp);
-                            bmp.Dispose();
-                            File.Delete(lstXml[i].uuid + ".png");
+                            List<CalculoNomina.Core.CodigoBidimensional> lstXml = new List<CalculoNomina.Core.CodigoBidimensional>();
                             try
                             {
                                 cnx.Open();
-                                nh.actualizaXml(GLOBALES.IDEMPRESA, dtpInicioPeriodo.Value.Date, dtpFinPeriodo.Value.Date, lstXml[i].idtrabajador, qr);
+                                lstXml = nh.obtenerListaQr(GLOBALES.IDEMPRESA, dtpInicioPeriodo.Value.Date, dtpFinPeriodo.Value.Date);
                                 cnx.Close();
                             }
-                            catch (Exception)
+                            catch (Exception error)
                             {
-                                MessageBox.Show("Error: Al actualizar el código QR.", "Error");
+                                MessageBox.Show("Error: Al obtener el listado de los XML." + error.Message, "Error");
                                 cnx.Dispose();
                                 return;
+                            }
+
+                            string codigoQR = "";
+                            string[] valores = null;
+                            string numero = "";
+                            string vEntero = "";
+                            string vDecimal = "";
+                            for (int i = 0; i < lstXml.Count; i++)
+                            {
+                                numero = lstXml[i].tt.ToString();
+                                valores = numero.Split('.');
+                                vEntero = valores[0];
+                                vDecimal = valores[1];
+                                codigoQR = string.Format("?re={0}&rr={1}&tt={2}.{3}&id={4}", lstXml[i].re, lstXml[i].rr,
+                                    vEntero.PadLeft(10, '0'), vDecimal.PadRight(6, '0'), lstXml[i].uuid);
+                                var qrEncoder = new QrEncoder(ErrorCorrectionLevel.H);
+                                var qrCode = qrEncoder.Encode(codigoQR);
+                                var renderer = new GraphicsRenderer(new FixedModuleSize(2, QuietZoneModules.Two), Brushes.Black, Brushes.White);
+
+                                using (var stream = new FileStream(lstXml[i].uuid + ".png", FileMode.Create))
+                                    renderer.WriteToStream(qrCode.Matrix, ImageFormat.Png, stream);
+
+                                Bitmap bmp = new Bitmap(lstXml[i].uuid + ".png");
+                                Byte[] qr = GLOBALES.IMAGEN_BYTES(bmp);
+                                bmp.Dispose();
+                                File.Delete(lstXml[i].uuid + ".png");
+                                try
+                                {
+                                    cnx.Open();
+                                    nh.actualizaXml(GLOBALES.IDEMPRESA, dtpInicioPeriodo.Value.Date, dtpFinPeriodo.Value.Date, lstXml[i].idtrabajador, qr);
+                                    cnx.Close();
+                                }
+                                catch (Exception)
+                                {
+                                    MessageBox.Show("Error: Al actualizar el código QR.", "Error");
+                                    cnx.Dispose();
+                                    return;
+                                }
                             }
                         }
                     }
